adding elements in postman

+++ src/main/java/com/cv/demo/db/Repository/ProjectsRep.java
@@ -2,7 +2,6 @@ package com.cv.demo.db.Repository;
 
 import com.cv.demo.db.Projects;
 import org.springframework.data.repository.CrudRepository;
-
 @org.springframework.stereotype.Repository
-public interface ProjectsRep extends CrudRepository<Projects , Long> {
+public interface ProjectsRep extends CrudRepository<Projects , Integer> {
 }
